\name{get.hwsd.con}
\alias{get.hwsd.con}
\title{get HWSD connection}
\usage{
  get.hwsd.con()
}
\value{
  connection to HWSD.sqlite database
}
\description{
  Function to create connection to HWSD.sqlite database
}
\details{
  copies database so that any changes are not saved in
  package
}
\author{
  David LeBauer
}

