\name{get.hwsdcon}
\alias{get.hwsdcon}
\title{get HWSD connection}
\usage{
  get.hwsdcon()
}
\value{
  connection to HWSD.sqlite database
}
\description{
  Function to create connection to HWSD.sqlite database
}
\details{
  copies database so that any changes are not saved in
  package
}
\author{
  David LeBauer
}

