\name{long2UTM}
\alias{long2UTM}
\title{long2UTM}
\usage{
  long2UTM(long)
}
\arguments{
  \item{long}{longitude}
}
\value{
  UTM Zone (integer)
}
\description{
  function to find a UTM zone
  \url{UTMzone}{http://en.wikipedia.org/wiki/Universal_Transverse_Mercator_coordinate_system#UTM_zone}
}
\author{
  D G Rossiter
}

