\name{myfun}
\alias{myfun}
\title{extract hwsd data from a region defined by a box}
\usage{
  myfun()
}
\arguments{
  \item{lat}{degrees latitude}

  \item{lon}{degrees longitude}

  \item{gridsize}{size of bounding box in degrees}
}
\value{
  records queried from region
}
\description{
  Function to extract and format one rectangular window
}
\author{
  D G Rossiter, David LeBauer
}

