% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwsdquery.R
\name{get.hwsd}
\alias{get.hwsd}
\title{extract hwsd data from a region}
\usage{
get.hwsd(x, con = con, hwsd.bil = NULL)
}
\arguments{
\item{x}{Either a vector specifying the corners of a box with \code{c(xmin, xmax, ymin, ymax)}; or a data frame with columns \code{lon} and \code{lat}.}

\item{con}{A connection to a data base returned by a function call \code{get.hwsd.con()}.}

\item{hwsd.bil}{(optional) location of file hwsd.bil. Too big for package repository. If necessary, this can be downloaded 
\url{here}{http://webarchive.iiasa.ac.at/Research/LUC/External-World-soil-database/HWSD_Data/HWSD_RASTER.zip}}
}
\value{
records queried from region
}
\description{
Function to extract and format one rectangular window
}
\examples{
\dontrun{
lat <- 44; lon <- -80; gridsize <- 0.1
x <- c(lon, lon, lat, lat) + gridsize/2 * c(-1, 1, -1, 1)
get.hwsd(x, con = con)
}
}
\author{
D G Rossiter, David LeBauer
}
